backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
Jinja2==3.1.2
psycopg2-binary==2.9.10
